+++ lib/consistency/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "version": "1.0.0",
   "description": "Strong eventual-consistency guarantees using CRDTs",
-  "exports": "./index.js",
+  "exports": "./dist/index.js",
   "scripts": {
     "build": "tsc",
     "dev": "tsc -w",
@@ -14,6 +14,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/node": "^24.3.0",
     "vitest": "^2.0.3"
   },
   "dependencies": {
