+++ lib/consistency/src/index.ts
