+++ lib/consistency/src/lww.ts
