+++ lib/consistency/src/map.test.ts
