+++ lib/pdf/package.json
@@ -4,7 +4,7 @@
   "description": "Export a list of incidents to a human-readable PDF report of the anomaly log",
   "main": "index.js",
   "type": "module",
-  "exports": "./index.js",
+  "exports": "./dist/index.js",
   "scripts": {
     "build": "tsc",
     "dev": "tsc -w"
