+++ lib/rules/src/main.ts
