+++ lib/rules/src/programs.ts
