--- lib/rules/worker.d.ts
@@ -1,1 +1,0 @@
-export type { Question } from "./qnaplus.d.ts";
