+++ lib/share/package.json
@@ -4,7 +4,10 @@
   "private": true,
   "version": "1.0.0",
   "description": "Share Instance Objects",
-  "exports": "./index.js",
+  "exports": {
+    "import": "./dist/index.js",
+    "types": "./dist/index.d.ts"
+  },
   "scripts": {
     "build": "tsc",
     "dev": "tsc -w",
@@ -14,6 +17,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/node": "^24.3.0",
     "vitest": "^2.0.3"
   },
   "dependencies": {
