+++ lib/share/src/asset.ts
