+++ lib/share/src/incident.ts
