+++ lib/share/src/index.ts
