+++ lib/share/src/scratchpad.ts
