+++ src/utils/hooks/qna.ts
@@ -7,7 +7,7 @@ import {
 import { ProgramAbbr, Year } from "robotevents";
 import { getQuestionsByProgram } from "~utils/data/qna";
 import { HookQueryOptions } from "./robotevents";
-import { relatedPrograms } from "@referee-fyi/rules/programs";
+import { relatedPrograms } from "../../../lib/rules/src/programs";
 
 export function getUseQuestionsForProgramQueryParams(
   program: ProgramAbbr,
